requirements.txt
# --- Core utilities ---
python-dotenv
requests
tiktoken

# --- Web app ---
streamlit

# --- LangChain ecosystem ---
langchain
langchain-core
langchain-openai
langchain-text-splitters
langchain-community

# --- LangGraph + checkpointing ---
langgraph
langgraph-prebuilt
langgraph-checkpoint-sqlite

# --- Vector store / PDFs / tools ---
faiss-cpu
pypdf
duckduckgo-search
tavily-python
langchain-tavily
ddgs

# --- OpenAI client (used under the hood by langchain-openai) ---
openai

# --- Pydantic (explicit, even though LangChain depends on it) ---
pydantic

# --- RAG-Anything / Multimodal RAG ---
raganything
lightrag-hku
Pillow
reportlab
unstructured  # For faster document parsing